refractor(ui): Split Link and Widget Component
